test(correlation): standardize remaining 4 API specs (custom-groups, identity-sets, reset, edge-cases)
Batch-convert journeyB/C/D/G onto the standard harness (enhanced-baseFixtures,
shared correlation-api-helpers, @correlation tags, per-test org + teardown,
Copyright+JSDoc, testLogger.testStart, standard naming). Assertions ported 1:1
(CUSTOM-01..03, IDENT-01..03, RESET-01/02, EDGE-01..03).

Remove the superseded journey*.api.spec.js files. All 5 API specs (14 tests)
collect cleanly under playwright.config.js.
